requirements.txt
uvicorn
pydantic-settings
